+++ formulario-colaborador/front-end/src/components/templates/Login/Login.css
@@ -13,7 +13,7 @@
     background-color: rgb(227, 227, 227);
     box-shadow: 0 2px 8px rgba(0,0,0,0.1);
     border-radius: 1.5rem;
-    height: 22rem;;
+    /* height: 22rem; */
 }
 
 p{
